refactor: 패키지 구조 변경
- 도메인 목적에 따라 패키지 분리
